+++ opentelemetry-datadog/src/lib.rs
@@ -42,13 +42,13 @@
 //!
 //! ```toml
 //! [dependencies]
-//! opentelemetry = { version = "*", features = ["rt-tokio"] }
+//! opentelemetry_sdk = { version = "*", features = ["rt-tokio"] }
 //! opentelemetry-datadog = "*"
 //! ```
 //!
 //! ```no_run
 //! # fn main() -> Result<(), opentelemetry::trace::TraceError> {
-//! let tracer = opentelemetry_datadog::new_pipeline()
+//! let provider = opentelemetry_datadog::new_pipeline()
 //!     .install_batch(opentelemetry_sdk::runtime::Tokio)?;
 //! # Ok(())
 //! # }
@@ -80,12 +80,12 @@
 //! [`DatadogPipelineBuilder`]: struct.DatadogPipelineBuilder.html
 //!
 //! ```no_run
-//! use opentelemetry::{KeyValue, trace::Tracer};
+//! use opentelemetry::{global, KeyValue, trace::{Tracer, TracerProvider}, InstrumentationScope};
 //! use opentelemetry_sdk::{trace::{self, RandomIdGenerator, Sampler}, Resource};
 //! use opentelemetry_sdk::export::trace::ExportResult;
-//! use opentelemetry::global::shutdown_tracer_provider;
 //! use opentelemetry_datadog::{new_pipeline, ApiVersion, Error};
 //! use opentelemetry_http::{HttpClient, HttpError};
+//! use opentelemetry_semantic_conventions as semcov;
 //! use async_trait::async_trait;
 //! use bytes::Bytes;
 //! use futures_util::io::AsyncReadExt as _;
@@ -115,7 +115,8 @@
 //! }
 //!
 //! fn main() -> Result<(), opentelemetry::trace::TraceError> {
-//!     let tracer = new_pipeline()
+//!     #[allow(deprecated)]
+//!     let provider = new_pipeline()
 //!         .with_service_name("my_app")
 //!         .with_api_version(ApiVersion::Version05)
 //!         .with_agent_endpoint("http://localhost:8126")
@@ -125,12 +126,20 @@
 //!                 .with_id_generator(RandomIdGenerator::default())
 //!         )
 //!         .install_batch(opentelemetry_sdk::runtime::Tokio)?;
+//!     global::set_tracer_provider(provider.clone());
+//!
+//!     let scope = InstrumentationScope::builder("opentelemetry-datadog")
+//!         .with_version(env!("CARGO_PKG_VERSION"))
+//!         .with_schema_url(semcov::SCHEMA_URL)
+//!         .with_attributes(None)
+//!         .build();
+//!     let tracer = provider.tracer_with_scope(scope);
 //!
 //!     tracer.in_span("doing_work", |cx| {
 //!         // Traced app logic here...
 //!     });
 //!
-//!     shutdown_tracer_provider(); // sending remaining spans before exit
+//!     provider.shutdown()?; // sending remaining spans before exit
 //!
 //!     Ok(())
 //! }
@@ -145,12 +154,12 @@ pub use exporter::{
 pub use propagator::{DatadogPropagator, DatadogTraceState, DatadogTraceStateBuilder};
 
 mod propagator {
-    use once_cell::sync::Lazy;
     use opentelemetry::{
         propagation::{text_map_propagator::FieldIter, Extractor, Injector, TextMapPropagator},
         trace::{SpanContext, SpanId, TraceContextExt, TraceFlags, TraceId, TraceState},
         Context,
     };
+    use std::sync::OnceLock;
 
     const DATADOG_TRACE_ID_HEADER: &str = "x-datadog-trace-id";
     const DATADOG_PARENT_ID_HEADER: &str = "x-datadog-parent-id";
@@ -163,13 +172,18 @@ mod propagator {
     const TRACE_STATE_TRUE_VALUE: &str = "1";
     const TRACE_STATE_FALSE_VALUE: &str = "0";
 
-    static DATADOG_HEADER_FIELDS: Lazy<[String; 3]> = Lazy::new(|| {
-        [
-            DATADOG_TRACE_ID_HEADER.to_string(),
-            DATADOG_PARENT_ID_HEADER.to_string(),
-            DATADOG_SAMPLING_PRIORITY_HEADER.to_string(),
-        ]
-    });
+    // TODO Replace this with LazyLock when MSRV is 1.80+
+    static TRACE_CONTEXT_HEADER_FIELDS: OnceLock<[String; 3]> = OnceLock::new();
+
+    fn trace_context_header_fields() -> &'static [String; 3] {
+        TRACE_CONTEXT_HEADER_FIELDS.get_or_init(|| {
+            [
+                DATADOG_TRACE_ID_HEADER.to_owned(),
+                DATADOG_PARENT_ID_HEADER.to_owned(),
+                DATADOG_SAMPLING_PRIORITY_HEADER.to_owned(),
+            ]
+        })
+    }
 
     #[derive(Default)]
     pub struct DatadogTraceStateBuilder {
@@ -449,7 +463,7 @@ mod propagator {
         }
 
         fn fields(&self) -> FieldIter<'_> {
-            FieldIter::new(DATADOG_HEADER_FIELDS.as_ref())
+            FieldIter::new(trace_context_header_fields())
         }
     }